feat: introduce AppShell sidebar + landing page restructure
Add a persistent sidebar navigation and split the monolithic dashboard
into a multi-tab architecture (refactor only — no new features or APIs).

- New route group app/(app)/ with a shared AppShell layout: desktop
  collapsible sidebar (persisted), mobile bottom nav, theme toggle + auth
  in the footer.
- AppShellContext centralises theme (isDark/primaryColor/themeMode) and
  wraps pages in the MUI ThemeProvider; reads/persists localStorage 'theme'.
- New landing page (/) — Market Overview: greeting + NYSE market status,
  MorningBriefingPanel, ticker search (routes to /analysis?symbol=),
  TrendingSparklines, SectorHeatmap. No prediction calls on load.
- Existing dashboard moved to /analysis; reads theme from context and
  auto-triggers analysis from the ?symbol= query param.
- Placeholder tabs for /options, /earnings, /ipo (Features 6-8).
- simulation back links switched to next/link.
