+++ frontend/app/(app)/earnings/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+// Placeholder — will be implemented in Feature 7
+import { Box, Typography } from '@mui/material';
+
+export default function EarningsPage() {
+  return (
+    <Box sx={{ p: 4 }}>
+      <Typography variant="h5">Earnings Calendar</Typography>
+      <Typography color="text.secondary">
+        Coming soon — upcoming earnings dates and estimates.
+      </Typography>
+    </Box>
+  );
+}
