+++ frontend/app/(app)/ipo/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+// Placeholder — will be implemented in Feature 8
+import { Box, Typography } from '@mui/material';
+
+export default function IpoPage() {
+  return (
+    <Box sx={{ p: 4 }}>
+      <Typography variant="h5">IPO Tracker</Typography>
+      <Typography color="text.secondary">
+        Coming soon — recent and upcoming initial public offerings.
+      </Typography>
+    </Box>
+  );
+}
