+++ frontend/app/(app)/options/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+// Placeholder — will be implemented in Feature 6
+import { Box, Typography } from '@mui/material';
+
+export default function OptionsPage() {
+  return (
+    <Box sx={{ p: 4 }}>
+      <Typography variant="h5">Options Chain</Typography>
+      <Typography color="text.secondary">
+        Coming soon — dedicated options chain with symbol search.
+      </Typography>
+    </Box>
+  );
+}
